Reset on network change
